feat(root): agent-swarm issue backlog + pre-swarm platform fixes
Development plan as 29 parallel-safe issues in docs/issues/ (protocol in
README, grab-list in INDEX, mermaid dependency graph), sized for one agent /
one PR each, TDD plans per issue mapped to SPEC $14.

Fixes landed alongside, so the swarm doesn't trip on them:
- api: autoload prefix mapping (routes/admin -> /admin/api), CORS regex that
  actually matches storefront origins, SPEC-shaped 429 body
- db: dbForShop deep-stamps nested writes via DMMF, scopes Shop/OrderSequence,
  memoizes per shop; 13 unit tests
- config: digit-wise fromDecimal (1.005 -> 101), sum() infers currency;
  11 unit tests; contracts: booleanish query params, int-only discount values
- theme: --theme-button-style emitted, footer-in-pages rejected
- tooling: start scripts load root .env, e2e/CI playwright + compose --wait
  fixes, verify runs env-parity, turbo globalEnv completed, single-source
  session cookie name

+++ packages/contracts/src/discounts.ts
@@ -46,8 +46,8 @@ export const discountSchema = z
     code: z.string().min(1).max(64).nullable().default(null),
     type: discountTypeSchema,
     valueType: discountValueTypeSchema,
-    /** percentage: 0–100. fixed: minor units. */
-    value: z.number().nonnegative(),
+    /** percentage: integer 0–100. fixed: integer minor units (SPEC §5 — never floats). */
+    value: z.number().int().nonnegative(),
     appliesTo: discountAppliesToSchema.default({ scope: 'all' }),
     minimumRequirement: minimumRequirementSchema.default({ type: 'none' }),
     usageLimit: z.number().int().positive().nullable().default(null),
